backend/services/requirements_clip.txt
flask==2.3.3
flask-cors==4.0.0
torch>=1.13.0
transformers>=4.21.0
pillow>=9.0.0
numpy>=1.21.0
python-dotenv>=0.19.0

